add fragment final ver
