+++ app/src/main/res/layout/fragment_transactions.xml
@@ -5,7 +5,8 @@
     <LinearLayout
         android:background="@color/backgroundlayout"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="wrap_content"
+        >
 
         <ListView
             android:id="@+id/list_view_id"
